Renamed FasterShots to Velocity and it can now be added to smasher
