+++ Tower Defence/Assets/Modules/Velocity/VelocityT2.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 8c62fa3adf3d14b4982b34740095f952, type: 3}
-  m_Name: FasterShotsT2
+  m_Name: VelocityT2
   m_EditorClassIdentifier: 
   moduleTier: 2
   accentColor: {r: 0.39215687, g: 0.5019608, b: 0.9411765, a: 1}
@@ -26,3 +26,6 @@ MonoBehaviour:
   bulletSpeedChange: 1
   rangeChange: 0.12
   fireRateChange: 0.12
+  smasherDamageChange: 0.15
+  smasherRangeChange: 0.15
+  smasherFireRateChange: 0.15
